requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
sqlalchemy
pandas
psycopg2-binary
